+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 ## TBD (UNRELEASED)
 
+FEATURES:
+* Official support for Zscaler Azure Tag Discovery Service
+* add: new examples ztags_standalone and module terraform-zscc-ztags-azure
+
 ENHANCEMENTS:
+* add: zsec support for Zscaler Azure Tag Discovery Service configuration
+* add: AzAPI provider (version 2.2.x) support for configuring Azure Event Grid resources
 * refactor: **support for future features** add base64 encoded user_data to VM and VMSS UserData attribute. CustomData will still be utilized until a future VHD release. At that time, it will remain as a secondary fallback to preferred UserData.
+
 
 ## v0.6.2 (December 9, 2024)